requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
nltk>=3.6.0
pyyaml>=5.4.0 